fix(server): MeetInfo::num_lifters is now a u16

+++ server/src/pages/meetlist.rs
@@ -102,7 +102,7 @@ pub struct MeetInfo<'db> {
     pub state: Option<&'db str>,
     pub town: Option<&'db str>,
     pub name: &'db str,
-    pub num_lifters: u32,
+    pub num_lifters: u16,
 }
 
 impl<'db> MeetInfo<'db> {
